requirements.txt
apache-airflow
openlineage-airflow
pandas
requests
psycopg2